feat: Integrate backend API services and implement authentication logic
- Implement `AuthRepository` and `AuthViewModel` to handle user login, registration, and session management using `TokenManager`.
- Implement `CourseRepository` and update `CourseViewModel` and `CourseDetailViewModel` to fetch real data from the backend API.
- Update data models (`User`, `Course`, `Enrollment`, `CourseMaterial`) to match backend schema with `SerializedName` annotations and `Int` identifiers.
- Configure `Retrofit` with `ApiConfig`, including an `authInterceptor` for JWT token handling and a logging interceptor.
- Implement the AI Chat functionality in `ChatViewModel` by connecting it to the backend RAG endpoint.
- Update `ProfileScreen` to include a login form and authenticated user profile view.
- Enable Internet permission and cleartext traffic (for local development) in `AndroidManifest.xml`.
- Refactor UI components and screens to handle nullability and use the new backend-driven data flow.

+++ MobileApp/app/src/main/java/com/elearning/ui/screens/ChatScreen.kt
@@ -1,6 +1,5 @@
 package com.elearning.ui.screens
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
@@ -14,7 +13,6 @@ import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import com.elearning.ui.data.model.ChatMessage
@@ -35,7 +33,7 @@ fun ChatScreen(
     val coroutineScope = rememberCoroutineScope()
 
     LaunchedEffect(courseId) {
-        viewModel.initializeChat(courseId)
+        viewModel.initializeChat(courseId.toInt())
     }
 
     LaunchedEffect(messages.size) {
